split out racing TestLogLevelParsing
